tr_scripts/requirements.txt
bpy
fake-bpy-module-3.4
numpy
open3d
Pillow
scipy
matplotlib
opencv-python